Use redis to cache metric pipeline to make sure vercel's vertical scaling doesn't break cache
- Moved cache implementation from blueprints to services.
- Updated URL shortener to use the new cache service.
- Increased cache timeout for metrics from 1 minute to 5 minutes.
- Removed unused imports and cleaned up code in URL shortener.
- Added Redis dependency to pyproject.toml for cache management.
