--- blueprints/cache.py
@@ -1,3 +1,0 @@
-from flask_caching import Cache
-
-cache = Cache(config={"CACHE_TYPE": "simple"})
