+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "pycountry>=24.6.1",
     "pymongo>=4.13.0",
     "python-dotenv>=1.1.0",
+    "redis>=6.2.0",
     "requests>=2.32.3",
     "tldextract>=5.3.0",
     "user-agents>=2.2.0",
